requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

pydantic==2.8.2
pydantic-settings==2.4.0

SQLAlchemy==2.0.32
asyncpg==0.29.0
alembic==1.13.2

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

redis==5.0.8
celery==5.4.0

email-validator